fix(calendar): fix calendar stories
