+++ core/components/organisms/calendar/__stories__/firstDayOfWeek.story.jsx
@@ -8,16 +8,14 @@ export const firstDayOfWeek = () => {
   return (
     <div className="d-flex flex-wrap">
       {values.map((v, index) => (
-        <div className="mr-9 mt-5" key={index}>
-          <Calendar date={new Date(2020, 2, 1)} firstDayOfWeek={v} />
-        </div>
+        <Calendar className="mr-9 mt-5" key={index} date={new Date(2023, 2, 1)} firstDayOfWeek={v} />
       ))}
     </div>
   );
 };
 
 export default {
-  title: 'Components/Calendar/Variants/First Day Of Week',
+  title: 'Components/Calendar/First Day Of Week',
   component: Calendar,
   parameters: {
     docs: {
