+++ core/components/organisms/calendar/__stories__/variants/size.story.jsx
@@ -7,12 +7,12 @@ export const size = () => {
     <div className="d-flex mt-8">
       <div className="mr-8">
         <Heading>size: small</Heading>
-        <Calendar date={new Date(2020, 2, 15)} size={'small'} view="date" />
+        <Calendar date={new Date(2023, 2, 15)} size={'small'} view="date" />
       </div>
 
       <div>
         <Heading>size: large</Heading>
-        <Calendar date={new Date(2020, 2, 15)} size={'large'} view="date" />
+        <Calendar date={new Date(2023, 2, 15)} size={'large'} view="date" />
       </div>
     </div>
   );
