+++ core/components/organisms/calendar/__stories__/variants/view.story.jsx
@@ -9,11 +9,9 @@ export const view = () => {
     <>
       <div className="mt-8">
         <Heading>size: small</Heading>
-        <div className="d-flex">
+        <div className="d-flex justify-content-between">
           {view.map((v, index) => (
-            <div className="mr-8" key={index}>
-              <Calendar date={new Date(2020, 2, 15)} size={'small'} view={v} />
-            </div>
+            <Calendar key={index} date={new Date(2023, 2, 15)} size={'small'} view={v} />
           ))}
         </div>
       </div>
@@ -22,9 +20,7 @@ export const view = () => {
         <Heading>size: large</Heading>
         <div className="d-flex">
           {view.map((v, index) => (
-            <div className="mr-8" key={index}>
-              <Calendar date={new Date(2020, 2, 15)} size={'large'} view={v} />
-            </div>
+            <Calendar key={index} date={new Date(2023, 2, 15)} size={'large'} view={v} />
           ))}
         </div>
       </div>
