+++ docs/src/pages/components/calendar/usage.mdx
@@ -94,4 +94,4 @@ Date picker is nothing but a calendar inside a popover and hence a trigger is al
 
 #### Date with an event
 Specific date can be highlighted to show that there is an event on that particular day.
-<Preview name="components-calendar-variants-date-with-events--date-with-events"/>
+<Preview name="components-calendar-date-with-events--date-with-events"/>
